Move recorder services to services.py (#71249)
