+++ .strict-typing
@@ -188,6 +188,7 @@ homeassistant.components.recorder.pool
 homeassistant.components.recorder.purge
 homeassistant.components.recorder.repack
 homeassistant.components.recorder.run_history
+homeassistant.components.recorder.services
 homeassistant.components.recorder.statistics
 homeassistant.components.recorder.system_health
 homeassistant.components.recorder.tasks
